+++ docs/pages/helpers/getcldimageurl/examples.mdx
@@ -35,6 +35,7 @@ import ImageGrid from '../../../components/ImageGrid';
       height="600"
       sizes="(max-width: 480px) 100vw, 50vw"
       alt=""
+      preserveTransformations
     />
 
     ### Basic
@@ -59,6 +60,7 @@ import ImageGrid from '../../../components/ImageGrid';
       height="600"
       sizes="(max-width: 480px) 100vw, 50vw"
       alt=""
+      preserveTransformations
     />
 
     ### Background Removal
