+++ docs/pages/helpers/getcldogimageurl/_meta.json
@@ -1,0 +1,5 @@
+{
+  "basic-usage": "Basic Usage",
+  "configuration": "Configuration",
+  "examples": "Examples"
+}
