+++ next-cloudinary/src/constants/sizes.ts
@@ -1,0 +1,2 @@
+export const OG_IMAGE_WIDTH = 2400;
+export const OG_IMAGE_HEIGHT = 1200;
