+++ next-cloudinary/src/index.ts
@@ -16,7 +16,10 @@ export type { CldVideoPlayerProps, CldVideoPlayerPropsColors, CldVideoPlayerProp
 export { cloudinaryLoader } from './loaders/cloudinary-loader';
 export type { CloudinaryLoader, CloudinaryLoaderLoaderOptions, CloudinaryLoaderCldOptions } from './loaders/cloudinary-loader';
 
-export { getCldImageUrl } from './lib/cloudinary';
-export type { GetCldImageUrl, GetCldImageUrlOptions, GetCldImageUrlConfig, GetCldImageUrlAnalytics } from './lib/cloudinary';
+export { getCldImageUrl } from './helpers/getCldImageUrl';
+export type { GetCldImageUrl, GetCldImageUrlOptions, GetCldImageUrlConfig, GetCldImageUrlAnalytics } from './helpers/getCldImageUrl';
+
+export { getCldOgImageUrl } from './helpers/getCldOgImageUrl';
+export type { GetCldOgImageUrl } from './helpers/getCldOgImageUrl';
 
 export type { CloudinaryVideoPlayer, CloudinaryVideoPlayerOptions, CloudinaryVideoPlayerOptionsColors, CloudinaryVideoPlayerOptionsLogo } from './types/player';
